jquery updates diffs

+++ themes_migrator.rb
@@ -29,10 +29,12 @@ Dir.glob(themes_root).each do |theme_root|
     else
       puts "Conflicts Found For: #{theme_root}"
 
-      file_path = File.join(theme_root, merge.path)
-      file = File.open("#{file_path}.diff", 'w+')
-      file.write(merge.result)
-      file.close
+      unless ENV['DRY_RUN']
+        file_path = File.join(theme_root, merge.path)
+        file = File.open("#{file_path}.diff", 'w+')
+        file.write(merge.result)
+        file.close
+      end
 
       file = File.open(File.join(theme_root, '..', 'conflicts.log'), 'a+')
       merge.conflicts.each do |type, details|
